feat: refactor confclient to use spring ctx (#2283)
* chore: refactor confclient to use spring ctx

* chore: remove unused rpc setup for now.
