+++ doc/Manuals/ug-syspar_x-road_v6_system_parameters.md
@@ -343,9 +343,9 @@ Proxy-ui has been removed in version 6.24 and it's parameters are not used anymo
 
 ### 3.6 Configuration Client parameters: `[configuration-client]`
 
-| **Parameter**                   | **Default value** | **Description**                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                |
+| **Parameter**                     | **Default value** | **Description**                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                |
 |-----------------------------------|-------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| port                              | 5665              | TCP port on which the configuration client process listens.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
+| port                              | 5665              | gRPC (TCP) port on which the configuration client process listens.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | update-interval                   | 60                | Global configuration download interval in seconds.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | admin-port                        | 5675              | TCP port on which the configuration client process listens for admin commands.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | allowed-federations               | none              | A comma-separated list of case-insensitive X-Road instances that fetching configuration anchors is allowed for. This enables federation with the listed instances if the X-Road instance is already federated at the central server level . Special value *none*, if present, disables all federation (the default value), while *all* allows all federations if *none* is not present. Example: *allowed-federations=ee,sv* allows federation with example instances *EE* and *Sv* while *allowed-federations=all,none* disables federation. X-Road services `xroad-confclient` and `xroad-proxy` need to be restarted (in that order) for the setting change to take effect. |
