+++ src/configuration-client/src/test/resources/configuration-anchor1.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tns:configurationAnchor xmlns:id="http://x-road.eu/xsd/identifiers"
+    xmlns:tns="http://x-road.eu/xsd/xroad.xsd"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://x-road.eu/xsd/xroad.xsd">
+
+    <instanceIdentifier>EE</instanceIdentifier>
+    <source>
+        <downloadURL>http://www.bar.com/conf</downloadURL>
+        <verificationCert>MIIDuzCCAqOgAwIBAgIITZrWd/qVYRYwDQYJKoZIhvcNAQEFBQAwNzERMA8GA1UE
+AwwIQWRtaW5DQTExFTATBgNVBAoMDEVKQkNBIFNhbXBsZTELMAkGA1UEBhMCU0Uw
+HhcNMTQwOTE2MjEzMTI2WhcNMTYwOTE1MjEzMTI2WjAyMQwwCgYDVQQGEwNBQUEx
+DDAKBgNVBAoMA0dPVjEUMBIGA1UEAwwLQUFBLWNlbnRyYWwwggEiMA0GCSqGSIb3
+DQEBAQUAA4IBDwAwggEKAoIBAQCCPKTL9CrQys9cIjZdB506rkUzRht0kWkGt46b
+kjQ/NlcjuRDLFGg23CguSUVolWGzbPm5mukhfx83oSJPd0T1R78zSm36sr2vWdsB
+HR9Jp9M3cnGIPF9TAKa6Sp6U0enqQUE9mAIqvvNBKGpIC2/sR0MjVWdOjN/uhzcS
+v2NAPUVAS5fQRvNexR07ynTdduzt9MHV9ddmkXYGxGhLXjfFE3kfqsbcRRzkMiCq
+iDwr6RF51NU+2uQAU8w3tUnZ1GdAhISvmAAQohyWOSv1uBObyhkEP3XQJmPPP6dK
+fo0ePblaWQ2yMPJrl1qeMc0mUgiF5tju8SjyLZFwQpam9n5pAgMBAAGjgc8wgcww
+TQYIKwYBBQUHAQEEQTA/MD0GCCsGAQUFBzABhjFodHRwOi8vbG9jYWxob3N0Ojgw
+ODAvZWpiY2EvcHVibGljd2ViL3N0YXR1cy9vY3NwMB0GA1UdDgQWBBSzRRtL5Ej2
+ZwkrQIhEMQha8LAu/DAMBgNVHRMBAf8EAjAAMB8GA1UdIwQYMBaAFHctiS4Dtv34
+0mU5MTUi0EYa6NIJMA4GA1UdDwEB/wQEAwIDuDAdBgNVHSUEFjAUBggrBgEFBQcD
+AQYIKwYBBQUHAwIwDQYJKoZIhvcNAQEFBQADggEBAAFOI5ry3EDCGInG76TzKyZk
+Ba49L8CxQQ7C+O/dnc4LzM7jn/a/WtlJzlsUoY3v+6c8sDcPxO8w3+vtnF3L55H6
+FFdWThM6ef4QajGpDshdkXGJbnbxmBDZ4KOutBMyxy0Afovowef8DIOnaZwidw9I
+aCuKTIA+rntRbxNaJlFVdG4u3yUqwg3hgwyyz+OYKE3y8OVGcAPsxU2aXbUyJKU1
+1fHnU+ICeTkhRbcD1Het00TE/BxObOGHSgXJOyr163vFSNzp8rVTs8BhYAV8XO2b
+sw/V7mNfhjoKJShw2hZCH+5DcnG+u6hz4iTyvTd1QRP61+wAJ7ueGxCMvA1fPmk=</verificationCert>
+<verificationCert>MIIDtjCCAp6gAwIBAgIIXoNLdbM9lM8wDQYJKoZIhvcNAQEFBQAwNzERMA8GA1UE
+AwwIQWRtaW5DQTExFTATBgNVBAoMDEVKQkNBIFNhbXBsZTELMAkGA1UEBhMCU0Uw
+HhcNMTQwOTE2MjE0MzM5WhcNMTYwOTE1MjE0MzM5WjAtMQwwCgYDVQQGEwNCQkIx
+DDAKBgNVBAoMA0NPTTEPMA0GA1UEAwwGeHJvYWQ2MIIBIjANBgkqhkiG9w0BAQEF
+AAOCAQ8AMIIBCgKCAQEAiX5n6ayod5GNo887htJvvFT80G7MVRoRE06LWsCCSisA
+IE4Pz4aMISHbIVUO6DJ05/GjaVrBHQXaezoLs1/VK+BA5n8J7fh/JyhqEzGALQmZ
+O/4y0bj1nXi80GHSh8Lx2fnoncRpWy/hYUiesmLM0Y76+Tnbm66GuuySl3dpUw9j
+wkIVp4faUxijyvF2x0djGY5tlBtZm22wyLrJTEzlDylv16RhAp7xjQLiuInupUPf
+jpNOZwXSAt42JAuEQDzgOkovdG7zEHDWXg2pkUmfWynlIiAQvN8eRkHfQk+HrXde
+bhR4spQGLYTWkg8E3E5rrMoLdNm/eu9pvEL8qRewdQIDAQABo4HPMIHMME0GCCsG
+AQUFBwEBBEEwPzA9BggrBgEFBQcwAYYxaHR0cDovL2xvY2FsaG9zdDo4MDgwL2Vq
+YmNhL3B1YmxpY3dlYi9zdGF0dXMvb2NzcDAdBgNVHQ4EFgQUu/LIP/15o5fkn71d
+7StEvk0A3T0wDAYDVR0TAQH/BAIwADAfBgNVHSMEGDAWgBR3LYkuA7b9+NJlOTE1
+ItBGGujSCTAOBgNVHQ8BAf8EBAMCA7gwHQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsG
+AQUFBwMCMA0GCSqGSIb3DQEBBQUAA4IBAQBFZAAFymlQ/mH+EB6pz3PxSeJHG3re
+Rj49EqTddfkcaq5cDYOzVytYd4D6c1niXVFbq88pFO7sXCH5dzfSSK2jg3QFEDYG
+zJ2fXXLTVzAaU1ZjDueCFOgUVee5JMrsgKDcsp5WJQhy/E3AhGdbWHbtIaZM70Ra
+NMUrRxurpqI+mlTSddrj1RRGwzBkzu3LiLtlz5v/T62BLqKqglKclhxNVxwYwnqW
+xCtXD6fwUZ9jQg0nbxdijZ/+GdWzmp41TOAKcDCMQVTQuo0m2A0gmg1aNPQp/SvR
+oYpF8VHcp7Ux8RMycbLu7YQF0sWk7XzochFdOanftap9JHe7/GSxTv3m</verificationCert>
+    </source>
+    <source>
+        <downloadURL>http://node-0x.example.com/internalconf</downloadURL>
+        <verificationCert>MIICqTCCAZGgAwIBAgIBATANBgkqhkiG9w0BAQ0FADAOMQwwCgYDVQQDDANOL0EwHhcNNzAwMTAxMDAwMDAwWhcNMzgwMTAxMDAwMDAwWjAOMQwwCgYDVQQDDANOL0EwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCAF9yFVF4cIkvC/uNidfktNy/EGo36SehcV4UsXLgnjM3qmVUtxIxc/PH9sHkPmH7tZSCmVf4iFwESqkp3aZlun7hedEni80Qtf0tHbSVsrj7kHU2MA9UAxnRG1r9SOVh8nkck/wpTGkOpxUqgNa6Xj6s38CYQPcQT+gaWZUoH1UOhbThI5oAKccB3YAWe7ghspVKyA0INqp9XPbgvuajuwbd8Vwfzqky8xiRSiZoeSirRLXy1Z3DcbPlJzF07l5nUHXeZp6+D6M7C6U4AWkytHJpg2Au4GigpLz9Y9zk1LoeF0vIpIkk49jpIKAQKL9Y2WAH94btB+lOrYEVHXytbAgMBAAGjEjAQMA4GA1UdDwEB/wQEAwIGQDANBgkqhkiG9w0BAQ0FAAOCAQEADbp4tIPNms/qi/24G4HkUyW+gjOm9t/bNPthQV8eEYxPfelIPKr5pBhooOLgiWIgB69afXjHG3dvBImwcjZAemxp5k/XCdH5DAkFEo/6NM31RwLL7NSouQAz9EdpQNfainhztRdLO/Lo4yuF8XSPfqdTFAek29fy7YC/f3bAHg2X6YMPj2szkbOo6zGuh58cRwPHRrfN7C+mlCdYS9tydBB+pOimPSZ4l/PPmGdncnh/DWVn9TzvTI/Ae8wV2I5g4/k3k83Ng836OWKq6PGbQhl1VnSt/dyNk9fI8wEslMGIpNxDMF+qB/O8AXN0T5A2ILQF+pDWA3H2VEBludo0+Q==
+        </verificationCert>
+    </source>
+    <source>
+        <downloadURL>http://node-1x.example.com/internalconf</downloadURL>
+        <verificationCert>MIICqTCCAZGgAwIBAgIBATANBgkqhkiG9w0BAQ0FADAOMQwwCgYDVQQDDANOL0EwHhcNNzAwMTAxMDAwMDAwWhcNMzgwMTAxMDAwMDAwWjAOMQwwCgYDVQQDDANOL0EwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCsgmCZg2SIApSlLx5mkGbeRapdcS5+8yE60D0un+N93qgBiDroJdLKlzCgBtcFXTX48yh/5krAP8xBH7jYrmSypFrZLQ7gjGw5CzYUeb7HZlj3k3A7qoEMueO4eR6nyZzqC2gs+CDo8WHSiXA0FaVQxHwv+EXwsaD2YAbwAPkY/n32kmUrs8O5+S3qlahWvVQb/Be6DOY/1xeO0MawF2lO+Fuy4bSZ8susbf8QiGRZ2LvlFXuqpnaH6nLY4Aaniu6xbt1idDJbo8WOabvEv6kmKpn5WAD4+MX9QIfxX7nXD6PQWINCmxYEUcJDGJ50BYpF+JKT10ynpKeVHCEQdAbPAgMBAAGjEjAQMA4GA1UdDwEB/wQEAwIGQDANBgkqhkiG9w0BAQ0FAAOCAQEAlFgW046DaMtrEarooGkCb04Ws0TTEQ7/ovI/L0tptXIGfaqaI5wDVc3rDt8X6LFCnariML3I9k7oOZc+irqDmQ88cupESZ49aJn9d1mfc2uURITOmRVAhRyYXp7y/Z0W0R4+XOnZ1mrDDFI8iDEY7vS2ce9iDvyg8KoAcjB8VZ1uZnMIw1m5r5z5Dkevn2dpFo60BLJH250HCZgNl/gk4Y8L7TExMaSyBI26/wHBMurEM7yUFTrc6OflFgBuiglip9KSaftI4LDm2oInt3x441gf0TyEjHXbPZ2sJ/E0ONKKLMxlieppa4xRSSLCF2b1db5RodJsy05ujBBxispsfQ==
+        </verificationCert>
+    </source>
+</tns:configurationAnchor>
