+++ src/configuration-client/src/test/resources/test-conf-simple/instance-identifier
@@ -1,0 +1,1 @@
+EE
