+++ src/configuration-proxy/scripts/download_instance_configuration.sh
@@ -14,5 +14,5 @@ CP="/usr/share/xroad/jlib/configuration-client.jar"
 
 XROAD_CONFCLIENT_PARAMS=" -Xmx50m -Dlogback.configurationFile=/etc/xroad/conf.d/confclient-logback.xml "
 
-java ${XROAD_PARAMS} ${XROAD_CONFCLIENT_PARAMS} -cp ${CP} ee.ria.xroad.common.conf.globalconf.ConfigurationClientMain $@
+java ${XROAD_PARAMS} ${XROAD_CONFCLIENT_PARAMS} -cp ${CP} org.niis.xroad.confclient.ConfClientCLIMain $@
