+++ src/packages/src/xroad/common/center/usr/share/xroad/scripts/verify_external_configuration.sh
@@ -19,5 +19,5 @@ XROAD_CONFCLIENT_PARAMS=" -Xmx50m -XX:MaxMetaspaceSize=70m \
 -Dxroad.appLog.xroad.level=$XROAD_LOG_LEVEL "
 
 
-java ${XROAD_PARAMS} ${XROAD_CONFCLIENT_PARAMS} -Dlogback.configurationFile=/etc/xroad/conf.d/confclient-logback.xml -cp /usr/share/xroad/jlib/configuration-client.jar  ee.ria.xroad.common.conf.globalconf.ConfigurationClientMain --verifyAnchorForExternalSource $@
+java ${XROAD_PARAMS} ${XROAD_CONFCLIENT_PARAMS} -Dlogback.configurationFile=/etc/xroad/conf.d/confclient-logback.xml -cp /usr/share/xroad/jlib/configuration-client.jar  org.niis.xroad.confclient.ConfClientCLIMain --verifyAnchorForExternalSource $@
